Change ClientError to extend from Exception instead of BaseException.
From the python documentation:  The built-in exception classes can be subclassed to define new exceptions; programmers are encouraged to derive new exceptions from the Exception class or one of its subclasses, and not from BaseException.

Having ClientError extend from BaseException causes problem when using this library in other tools (such as the behave framework) that consider exceptions that extend directly from BaseException to be terminal and exit the program.